RR-358 - Move messaging related code to package of that name
